Phase B-2: kx-tool-registry lib.rs split (Rule 3, behavior-preserving move-only)
Second Phase B refactor per Rule 3 (no giant lib.rs). Pure move; zero
behavior change.

Before: src/lib.rs 1504 lines (ID newtypes + ToolKind + IdempotencyClass
+ ToolDef + ToolResolutionEvent + ResolvedTool + ToolProvenance +
RegistrationStatus + 2 error enums + registration_token_of helper +
ToolRegistry trait + InMemoryToolRegistry + 17 inline unit tests, all in
one file).

After: thin src/lib.rs 105 lines (crate attrs + crate doc + mod decls +
pub use re-exports + #[cfg(test)] mod tests; + the existing
kx_warrant::{FsScope, NetScope, ResourceCeiling} re-export) + 9 sibling
modules organized by concern:

- ids.rs (3 newtype IDs: McpEndpointId, ReviewerId, RegistrationToken)
- tool_kind.rs (the ToolKind enum)
- idempotency_class.rs (IdempotencyClass — Seam B per D38 §2)
- tool_def.rs (ToolDef + ToolResolutionEvent + impl + ResolvedTool)
- provenance.rs (ToolProvenance + RegistrationStatus)
- errors.rs (ResolutionError + RegistrationError)
- token.rs (registration_token_of helper)
- registry.rs (ToolRegistry trait + private RegistrationRecord +
  InMemoryToolRegistry incl. with_builtins())
- tests.rs (the 17 inline unit tests; #[cfg(test)] from lib.rs)

Re-exports preserve every previously-public path — kx_tool_registry::{
McpEndpointId, RegistrationToken, ReviewerId, ToolKind, IdempotencyClass,
ToolDef, ToolResolutionEvent, ResolvedTool, ToolProvenance,
RegistrationStatus, ResolutionError, RegistrationError,
registration_token_of, ToolRegistry, InMemoryToolRegistry, FsScope,
NetScope, ResourceCeiling} all unchanged. No downstream crate's use
statements needed editing (cargo build --workspace clean).

Local gates green on Apple Silicon (--all-features):
- cargo build -p kx-tool-registry: clean
- cargo test --workspace --all-features: 645 + 8 (IDENTICAL to pre-split)
- cargo test -p kx-tool-registry: 20 + 3 + 8 + 6 = 37 (IDENTICAL)
- cargo clippy -p kx-tool-registry --all-targets -- -D warnings: clean
- cargo fmt --all -- --check: clean

Next Phase B slice: kx-warrant split (1254 lines).
